pyproject.toml
[project]
name = "ndp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py==2.1.0",
    "aiobotocore==2.16.0",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aioitertools==0.12.0",
    "aiosignal==1.3.2",
    "albucore==0.0.21",
    "albumentations==1.4.23",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.7.0",
    "arrow==1.3.0",
    "async-timeout==5.0.1",
    "attrs==24.3.0",
    "beautifulsoup4==4.12.3",
    "blessed==1.20.0",
    "boto3==1.35.81",
    "botocore==1.35.81",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "croniter==1.3.15",
    "datasets==3.2.0",
    "dateutils==0.6.12",
    "deepdiff==7.0.1",
    "dill==0.3.8",
    "docker-pycreds==0.4.0",
    "docstring-parser==0.16",
    "editor==1.6.6",
    "einops==0.8.0",
    "eval-type-backport==0.2.0",
    "exceptiongroup==1.2.2",
    "filelock==3.16.1",
    "fire==0.7.0",
    "frozenlist==1.5.0",
    "fsspec==2023.12.2",
    "future==1.0.0",
    "gitdb==4.0.11",
    "gitpython==3.1.43",
    "grpcio==1.68.1",
    "h11==0.14.0",
    "huggingface-hub==0.27.0",
    "idna==3.10",
    "imageio==2.36.1",
    "imageio-ffmpeg==0.5.1",
    "importlib-metadata==8.5.0",
    "importlib-resources==6.4.5",
    "inquirer==3.4.0",
    "ipykernel>=6.29.5",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.4",
    "jmespath==1.0.1",
    "jsonargparse==4.35.0",
    "lightning==2.5.0.post0",
    "lightning-cloud==0.5.70",
    "lightning-lite==1.8.6",
    "lightning-utilities==0.11.9",
    "markdown==3.7",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib>=3.10.0",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "multiprocess==0.70.16",
    "muon",
    "networkx==3.4.2",
    "numpy==2.2.0",
    "omegaconf==2.3.0",
    "opencv-python==4.10.0.84",
    "opencv-python-headless==4.10.0.84",
    "ordered-set==4.1.0",
    "packaging==22.0",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pillow==11.0.0",
    "platformdirs==4.3.6",
    "propcache==0.2.1",
    "protobuf==5.29.2",
    "psutil==6.1.1",
    "pudb==2024.1.3",
    "pyarrow==18.1.0",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "pygments==2.18.0",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-multipart==0.0.20",
    "pytorch-lightning==2.5.0.post0",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "readchar==4.2.1",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==13.9.4",
    "runs==1.2.2",
    "s3fs==2023.12.2",
    "s3transfer==0.10.4",
    "safetensors==0.4.5",
    "scipy==1.14.1",
    "sentry-sdk==2.19.2",
    "setproctitle==1.3.4",
    "simsimd==6.2.1",
    "six==1.17.0",
    "smmap==5.0.1",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "starlette==0.22.0",
    "starsessions==1.3.0",
    "stringzilla==3.11.2",
    "sympy==1.13.1",
    "tensorboard==2.18.0",
    "tensorboard-data-server==0.7.2",
    "termcolor==2.5.0",
    "test-tube==0.7.5",
    "tokenizers==0.21.0",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchmetrics==0.11.4",
    "torchvision>=0.20.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformer-engine-cu12==1.11.0",
    "transformers==4.47.1",
    "types-python-dateutil==2.9.0.20241206",
    "typeshed-client==2.7.0",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "urwid==2.6.16",
    "urwid-readline==0.15.1",
    "uvicorn==0.34.0",
    "wandb==0.19.1",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==11.0.3",
    "werkzeug==3.1.3",
    "wrapt==1.17.0",
    "xmod==1.8.1",
    "xxhash==3.5.0",
    "yarl==1.18.3",
    "zipp==3.21.0",
]

[tool.uv.sources]
muon = { git = "https://github.com/KellerJordan/Muon" }
